fix: make it a reusable plugin so that it can be colocated with nodes eligible for placeholders
