+++ shared/editor/nodes/index.ts
@@ -1,7 +1,6 @@
 import DateTime from "../extensions/DateTime";
 import History from "../extensions/History";
 import MaxLength from "../extensions/MaxLength";
-import Placeholder from "../extensions/Placeholder";
 import TrailingNode from "../extensions/TrailingNode";
 import Extension from "../lib/Extension";
 import Bold from "../marks/Bold";
@@ -64,7 +63,6 @@ export const basicExtensions: Nodes = [
   Strikethrough,
   History,
   TrailingNode,
-  Placeholder,
   MaxLength,
   DateTime,
 ];
